Revert "Controller Utility Class"
This reverts commit ad94440c037d4a86877234a8e3dd7f8aaeab3acf.
